Renamed FullStatsModals to FullStatsModal
